requirements.txt
alpaca-py==0.20.3
pandas==2.2.3
numpy==2.1.3
python-dotenv==1.0.1
pytz==2024.2
pyyaml==6.0.2
